Add integration tests to remove database tests from acceptance tests
Acceptance tests are run against the staging environment, which simulates a productive environment and thus gives no access to the database directly. To test if value are stored correctly in the database integration tests (which are basically the same as acceptance tests + database tests) are run against a development environment, which gives direct access to the database.